+++ pnpm-workspace.yaml
@@ -20,7 +20,7 @@ catalog:
   '@mui/icons-material': ^7.3.7
   '@mui/internal-test-utils': 2.0.18-canary.12
   '@mui/material': ^7.3.7
-  '@mui/stylis-plugin-rtl': '9.0.0-alpha.2'
+  '@mui/stylis-plugin-rtl': '9.0.0-alpha.3'
   '@mui/system': ^7.3.7
   '@mui/utils': ^7.3.7
   '@mui/x-internal-exceljs-fork': 4.4.4
